Contest time utilities
Move contest time functions to separate file and expose as individual functions.
Besides cleaner separation, this allows more control over what we expose and some
simplification in use.
